Extracted Paintable class to its own file
